Revert "[#71] Explicit schema validation"
This reverts commit 588ea534065f422219bbe822a24c40f071e66a09.
