+++ liquigraph-core/src/main/java/org/liquigraph/core/io/xml/SchemaErrorHandler.java
@@ -22,6 +22,7 @@ import org.xml.sax.ErrorHandler;
 import org.xml.sax.SAXException;
 import org.xml.sax.SAXParseException;
 
+import java.util.ArrayList;
 import java.util.Collection;
 
 /**
