+++ liquigraph-core/src/main/java/org/liquigraph/core/io/xml/XmlSchemaValidator.java
@@ -16,13 +16,31 @@
 package org.liquigraph.core.io.xml;
 
 import org.w3c.dom.Node;
+import org.xml.sax.SAXException;
 
+import javax.xml.XMLConstants;
 import javax.xml.transform.dom.DOMSource;
+import javax.xml.transform.stream.StreamSource;
+import javax.xml.validation.Schema;
+import javax.xml.validation.SchemaFactory;
+import javax.xml.validation.Validator;
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.Collection;
 
 import static java.lang.String.format;
 
 public class XmlSchemaValidator {
+
+    private final Schema schema;
+
+    public XmlSchemaValidator() {
+        try (InputStream stream = getClass().getResourceAsStream("/schema/changelog.xsd")) {
+            schema = SchemaFactory.newInstance(XMLConstants.W3C_XML_SCHEMA_NS_URI).newSchema(new StreamSource(stream));
+        } catch (SAXException | IOException e) {
+            throw new IllegalStateException(e);
+        }
+    }
 
     /**
      * Validate the fully resolved changelog containing all migrations
@@ -37,8 +55,10 @@ public class XmlSchemaValidator {
 
     private Collection<String> validate(DOMSource changelog) {
         try {
-            return parse(changelog);
-        } catch (Exception e) {
+            SchemaErrorHandler customErrorHandler = new SchemaErrorHandler();
+            validator(customErrorHandler).validate(changelog);
+            return customErrorHandler.getErrors();
+        } catch (IOException | SAXException e) {
             throw new IllegalArgumentException(
                 format("Exception while reading changelog: %n\t%s.", e.getMessage()),
                 e
@@ -46,10 +66,10 @@ public class XmlSchemaValidator {
         }
     }
 
-    private Collection<String> parse(DOMSource changelog) throws Exception {
-        return new DomSourceValidatorFactory()
-                .createValidator(changelog)
-                .validate(changelog);
+    private Validator validator(SchemaErrorHandler customErrorHandler) {
+        Validator validator = schema.newValidator();
+        validator.setErrorHandler(customErrorHandler);
+        return validator;
     }
 
 }
